chore(express example): files organisation
