+++ test/examples/wrappers/express/plan/make-adapters.js
@@ -1,0 +1,4 @@
+const GiveTheSecretHTTPAdapter = require('../technical/http/adapters/give-the-secret-http-adapter');
+
+module.exports = (context) => context
+  .addUsingClass('expressAdapter', GiveTheSecretHTTPAdapter);
