+++ test/examples/wrappers/express/plan/make-business.js
@@ -1,0 +1,3 @@
+module.exports = (context) => context
+  .addFunction('ensureAuthenticated', () => {})
+  .addValue('secret', 'I love you');
