+++ test/examples/wrappers/express/plan/make-libs.js
@@ -1,0 +1,4 @@
+/* eslint-disable global-require */
+module.exports = (context) => context
+  .addInstance('http', require('http'))
+  .addInstance('express', require('express'));
